+++ __tests__/Home.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import '@testing-library/jest-dom'
 import { MemoryRouter, Route, Routes } from 'react-router-dom'
-import Home from '../src/pages/Home'
+import Home from '../src/pages/Home/Home'
 import { fireEvent, render, cleanup } from '@testing-library/react'
 
 let currentUser = { id: 1, username: "lincolnpepper" }
